refactor: 키친라이더는 주문 기능에서 사용하므로 주문 아래 common 패키지로 이동.
