+++ src/main/java/kitchenpos/order/common/KitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.order.common;
 
 import java.math.BigDecimal;
 import java.util.UUID;
